requirements.txt
setuptools
pyautoit
pypiwin32
pywin32
PyAutoGUI